Lift the state up from Add To App
- Introduce a new generic service file that is responsible for
fetching/sending requests to the backend.

- Remove fetching products types from Add Product component to the
`App`  component instead so that the testing bot
can find the dropdown options ready before mounting the `AddProduct`
Page as it wasn't able before due to the async fetching that runs after
rendering!

- Get rid of the uneeded `ref` checkers of the `useEffect` hooks, since that rendering
won't occur twice except for the development env. On Production
it's safe (So Far).